requirements.txt
streamlit
numpy
scipy
matplotlib
pillow
soundfile


